Add tests for sending commands and statuses
